// reactapp/src/models/enum/NamesInputs.ts
export enum NamesInputs {
    INN = "INN",
    ScanINN = "ScanINN",
    OGRNIP = "OGRNIP",
    ScanOGRNIP = "ScanOGRNIP",
    DateOfRegistration = "DateOfRegistration",
    ScanOfAnExtractFromTheUSRIP = "ScanOfAnExtractFromTheUSRIP",
    ScanLeaseAgreementPremises = "ScanLeaseAgreementPremises",
    ThereNoContract = "ThereNoContract",
    Fullname = "Fullname",
    ShortName = "ShortName",
    OGRN = "OGRN",
    ScanOGRN = "ScanOGRN",
    BIC = "BIC",
    NameBankBranch = "NameBankBranch",
    CorrespondentAccount = "CorrespondentAccount",
    PaymentAccount = "PaymentAccount",
}